Add pushback, more groundwork, see detailed msg Implemented pushback as seen in the java version. Matt and I realized it may not be entirely correct, so more changes to come.
Changed various struct types to classes so they can have constructors
and generally conform to object-oriented design.

Determined a datatype for vdata. It uses a vector, but the size of this
vector should never be changed, only set once at initialization. I've
also swapped to using atomic integers instead of references that are
atomic. The minimum value of an integer shall be reserved for marking
as logically deleted.

Removed typedef from enum b/c that is not required in C++.

The groundwork was creating method skeletons for the many other
methods that need implementing, though we still need to make
constructors for many of the classes.